refactor: OrderTable 리팩토링 - OrderTableService의 작업 중 동일한 애그리거트 작업을 도메인 작업으로 캡슐화
