+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -58,8 +58,8 @@ public class TableGroupService {
 
         final Long tableGroupId = savedTableGroup.getId();
         for (final OrderTable savedOrderTable : savedOrderTables) {
-            savedOrderTable.setTableGroupId(tableGroupId);
             savedOrderTable.setEmpty(false);
+            savedOrderTable.setTableGroupId(tableGroupId);
             orderTableDao.save(savedOrderTable);
         }
         savedTableGroup.setOrderTables(savedOrderTables);
