+++ src/main/java/kitchenpos/dao/OrderTableDao.java
@@ -5,6 +5,10 @@ import kitchenpos.domain.OrderTable;
 import org.springframework.data.repository.CrudRepository;
 
 public interface OrderTableDao extends CrudRepository<OrderTable, Long> {
+    default OrderTable findMandatoryById(Long id) {
+        return findById(id).orElseThrow(IllegalArgumentException::new);
+    }
+
     List<OrderTable> findAll();
 
     List<OrderTable> findAllByIdIn(List<Long> ids);
